zb: ObjectServer keeps owned Connection
This is achieved w/o breaking the API by making use of the new Clone
implementation of Connection that's cheap. We also add a
`new_own_connection` constructor so users can explicitly pass the
ownership of the connection to ObjectServer.

We do remove the lifetimes from `ObjectServer` struct but that's not an
API break as all existing code will continue to build and work.

Fixes https://gitlab.freedesktop.org/zeenix/zbus/-/issues/58