wip: OwnedProxy type
An experiment on having a separate type for owned Proxy that derefs to
Proxy. The issue is the DBusProxy macro: It'd make sense to generate two
structs (owned and unowed) for each proxy but what should the owned
proxy struct keep? The `OwnedProxy` struct or the generated unowned `MyProxy`?

If we keep `OwnedProxy`, we can't deref to `MyProxy`, which beats the
point of the exercise. If we keep `MyProxy`, how do the constructors
create an owned version of `MyProxy`?