zb: Drop use of TLS for Connection
TLS complicates code and I'd want to not use them at all but it's not
possible to drop the LOCAL_NODE TLS w/o breaking API. However,
Connection is very much droppable by keeping associated connection in the
ObjectNode itself. Now that Connection is cheap to clone, this change
doesn't introduce any runtime/allocation costs.

We also deprecate the Error::NoTLSConnection as we won't be returing
that ever anymore.